Order recompile Online

+++ apps/sms_ussd/models.py
@@ -22,10 +22,10 @@ class SIM(models.Model, ):
                                 max_length=14,
                                 null=True,
                                 blank=True, )
-    imsi = models.BigIntegerField(verbose_name=_(u'IMSI', ),
-                                  unique=True,
-                                  null=False,
-                                  blank=False, )
+    imsi = models.IntegerField(verbose_name=_(u'IMSI', ),
+                               unique=True,
+                               null=False,
+                               blank=False, )
 
     #Дата создания и дата обновления. Устанавливаются автоматически.
     created_at = models.DateTimeField(auto_now_add=True,
